:recycle: refactor: update struct of project
