+++ backend/app/main.py
@@ -1,4 +1,4 @@
-from .config import create_app
+from app import create_app
 
 app = create_app()
